Split jupyter_collaborative_chat fontend package and python extension
