+++ pyproject.toml
@@ -37,7 +37,7 @@ fields = ["description", "authors", "urls"]
 [tool.jupyter-releaser.options]
 version-cmd = "cd ../.. && python scripts/bump_version.py --force --skip-if-dirty"
 python_packages = [
-    "packages/jupyterlab-collaborative-chat",
+    "python/jupyterlab-collaborative-chat",
     "packages/jupyterlab-ws-chat"
 ]
