+++ python/jupyterlab-collaborative-chat/.prettierignore
