+++ python/jupyterlab-collaborative-chat/install.json
