+++ python/jupyterlab-collaborative-chat/jupyter-config/server-config/jupyterlab_collaborative_chat.json
