+++ python/jupyterlab-collaborative-chat/jupyterlab_collaborative_chat/__init__.py
@@ -15,7 +15,7 @@ except ImportError:
 def _jupyter_labextension_paths():
     return [{
         "src": "labextension",
-        "dest": "jupyterlab-collaborative-chat"
+        "dest": "jupyterlab-collaborative-chat-extension"
     }]
