+++ python/jupyterlab-collaborative-chat/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyter/chat",
+  "name": "jupyterlab-collaborative-chat-extension",
   "version": "0.3.1",
-  "description": "A package that provides UI components that can be used to create a chat in a Jupyterlab extension.",
+  "description": "A chat extension based on shared documents",
   "keywords": [
     "jupyter",
     "jupyterlab",
@@ -30,14 +30,17 @@
     "url": "https://github.com/jupyterlab/jupyter-chat.git"
   },
   "scripts": {
-    "build": "jlpm build:lib",
-    "build:prod": "jlpm clean && jlpm build:lib:prod",
+    "build": "jlpm build:lib && jlpm build:labextension:dev",
+    "build:prod": "jlpm clean && jlpm build:lib:prod && jlpm build:labextension",
+    "build:labextension": "jupyter labextension build .",
+    "build:labextension:dev": "jupyter labextension build --development True .",
     "build:lib": "tsc --sourceMap",
     "build:lib:prod": "tsc",
     "clean": "jlpm clean:lib",
     "clean:lib": "rimraf lib tsconfig.tsbuildinfo",
     "clean:lintcache": "rimraf .eslintcache .stylelintcache",
-    "clean:all": "jlpm clean:lib && jlpm clean:lintcache",
+    "clean:labextension": "rimraf jupyterlab_collaborative_chat/labextension jupyterlab_collaborative_chat/_version.py",
+    "clean:all": "jlpm clean:lib && jlpm clean:labextension && jlpm clean:lintcache",
     "eslint": "jlpm eslint:check --fix",
     "eslint:check": "eslint . --cache --ext .ts,.tsx",
     "install:extension": "jlpm build",
@@ -49,32 +52,41 @@
     "stylelint": "jlpm stylelint:check --fix",
     "stylelint:check": "stylelint --cache \"style/**/*.css\"",
     "test": "jest --coverage",
-    "watch:src": "tsc -w --sourceMap"
+    "watch": "run-p watch:src watch:labextension",
+    "watch:src": "tsc -w --sourceMap",
+    "watch:labextension": "jupyter labextension watch ."
   },
   "dependencies": {
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@jupyter/react-components": "^0.15.2",
+    "@jupyter/collaboration": "^2.1.0",
+    "@jupyter/docprovider": "^2.1.0",
+    "@jupyter/ydoc": "^1.1.1",
     "@jupyterlab/application": "^4.2.0",
     "@jupyterlab/apputils": "^4.3.0",
+    "@jupyterlab/coreutils": "^6.2.0",
+    "@jupyterlab/docregistry": "^4.2.0",
+    "@jupyterlab/launcher": "^4.2.0",
     "@jupyterlab/notebook": "^4.2.0",
     "@jupyterlab/rendermime": "^4.2.0",
+    "@jupyterlab/services": "^7.2.0",
+    "@jupyterlab/settingregistry": "^4.2.0",
+    "@jupyterlab/translation": "^4.2.0",
     "@jupyterlab/ui-components": "^4.2.0",
     "@lumino/commands": "^2.0.0",
-    "@lumino/disposable": "^2.0.0",
+    "@lumino/coreutils": "^2.0.0",
     "@lumino/signaling": "^2.0.0",
-    "@mui/icons-material": "^5.11.0",
-    "@mui/material": "^5.11.0",
-    "clsx": "^2.1.0",
+    "@lumino/widgets": "^2.0.0",
+    "jupyterlab-collaborative-chat": "^0.3.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "y-protocols": "^1.0.5",
+    "yjs": "^13.5.40"
   },
   "devDependencies": {
+    "@jupyterlab/builder": "^4.2.0",
+    "@jupyterlab/testutils": "^4.2.0",
     "@types/jest": "^29.2.0",
     "@types/json-schema": "^7.0.11",
     "@types/react": "^18.2.0",
     "@types/react-addons-linked-state-mixin": "^0.14.22",
-    "@types/react-dom": "^18.2.0",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
     "css-loader": "^6.7.1",
@@ -82,6 +94,7 @@
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^5.0.0",
     "jest": "^29.2.0",
+    "mkdirp": "^1.0.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.0",
     "rimraf": "^5.0.1",
@@ -92,7 +105,8 @@
     "stylelint-config-standard": "^34.0.0",
     "stylelint-csstree-validator": "^3.0.0",
     "stylelint-prettier": "^4.0.0",
-    "typescript": "~5.0.2"
+    "typescript": "~5.0.2",
+    "yjs": "^13.5.0"
   },
   "sideEffects": [
     "style/*.css",
@@ -101,6 +115,31 @@
   "styleModule": "style/index.js",
   "publishConfig": {
     "access": "public"
+  },
+  "jupyterlab": {
+    "discovery": {
+      "server": {
+        "managers": [
+          "pip"
+        ],
+        "base": {
+          "name": "jupyterlab_collaborative_chat"
+        }
+      }
+    },
+    "extension": true,
+    "outputDir": "jupyterlab_collaborative_chat/labextension",
+    "schemaDir": "schema",
+    "sharedPackages": {
+      "@jupyter/docprovider": {
+        "bundled": true,
+        "singleton": true
+      },
+      "jupyterlab-collaborative-chat": {
+        "bundled": true,
+        "singleton": true
+      }
+    }
   },
   "eslintIgnore": [
     "node_modules",
