+++ python/jupyterlab-collaborative-chat/schema/chat-panel.json
