+++ python/jupyterlab-collaborative-chat/schema/commands.json
