+++ python/jupyterlab-collaborative-chat/schema/factory.json
