+++ python/jupyterlab-collaborative-chat/screenshot.gif
