+++ python/jupyterlab-collaborative-chat/style/base.css
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) Jupyter Development Team.
+ * Distributed under the terms of the Modified BSD License.
+ */
+
+/*
+  See the JupyterLab Developer Guide for useful CSS Patterns:
+
+  https://jupyterlab.readthedocs.io/en/stable/developer/css.html
+*/
+
+@import url('~jupyterlab-collaborative-chat/style/index.css');
+
