+++ python/jupyterlab-collaborative-chat/style/index.js
@@ -1,0 +1,6 @@
+/*
+ * Copyright (c) Jupyter Development Team.
+ * Distributed under the terms of the Modified BSD License.
+ */
+
+import './base.css';
