+++ python/jupyterlab-collaborative-chat/ui-tests/jupyter_server_test_config.py
