+++ python/jupyterlab-collaborative-chat/ui-tests/test-results/.last-run.json
@@ -1,0 +1,7 @@
+{
+  "status": "failed",
+  "failedTests": [
+    "86659aada0458a2dec90-ed575443018384c7f71c",
+    "e45d9e38adc663a001d3-2a5641d1d85d64c528b1"
+  ]
+}
