+++ python/jupyterlab-collaborative-chat/ui-tests/tests/autocompletion.spec.ts
