+++ scripts/install.sh
@@ -11,12 +11,12 @@ jlpm install
 if [ -z "${EXTENSION}" ]; then
   jlpm build
   # install the collaborative chat extension
-  pip install -e packages/jupyterlab-collaborative-chat[test]
+  pip install -e python/jupyterlab-collaborative-chat[test]
 
   # install websocket chat extension
   pip install -e packages/jupyterlab-ws-chat[test]
 else
   PACKAGE="jupyterlab-${EXTENSION}-chat"
   jlpm build:${EXTENSION}
-  pip install -e packages/${PACKAGE}[test]
+  pip install -e python/${PACKAGE}[test]
 fi
